Ændret nogle ints til double

+++ Repetition Inheritance/BenzinBil.cs
@@ -14,7 +14,7 @@ namespace Repetition_Inheritance
         }
 
         public int Tank { get; set; }
-        public override int HalvÅrligEjerafgift()
+        public override double HalvÅrligEjerafgift()
         {
             int ejerafgft = 0;
             if (KmPrLiter < 20)
@@ -32,7 +32,7 @@ namespace Repetition_Inheritance
             return ejerafgft;
         }
 
-        public override int Rækkevidde()
+        public override double Rækkevidde()
         {
             int række = Tank*KmPrLiter;
             return række;
